==> Ekimetrics/06d. Total Database - Joins - Blank3.ipynb <==
## Read in Libraries

from datetime import datetime
import pandas as pd
import numpy as np
import warnings

warnings.filterwarnings('ignore') 

## Define import path for all treated pieces of data and export path

# This Python code assigns file paths to three data files:
# Facebook Data: path_fb points to "final_facebook.csv" in the "Processed Data" directory.
# YouTube Data: path_yt points to "final_youtube.csv" in the "Processed Data" directory.
# Outdoor Data: path_outdoor points to "final_outdoor.csv" in the "Processed Data" directory.
# These file locations will be used to retrieve the processed data files later on, either for further analysis or inclusion in a final database.

path_fb = r'C:/Users/drvis/Downloads/MMM-2/Processed Data/final_facebook.csv'

path_yt = r'C:/Users/drvis/Downloads/MMM-2/Processed Data/final_youtube.csv'

path_outdoor = r'C:/Users/drvis/Downloads/MMM-2/Processed Data/final_outdoor.csv'

# STUDENT INPUT REQUIRED - Modify path directly below for location of the 06a. All Other Variables.csv file on your laptop/desktop
path_other_variables = r'C:/Users/drvis/Downloads/MMM-2/06a. All Other Variables (1).csv'

export_data_path = 'C:/Users/drvis/Downloads/MMM-2/Processed Data/'
file_name = 'final_database.csv'

## Read in Treated Facebook/Instagram, YouTube, and Outdoor Campaign Data

# This Python code is intended to read in the processed data from CSV files into Pandas DataFrames. It reads four separate CSV files:
# df_fb: Loads the processed Facebook data from the file specified in `path_fb` into a DataFrame with the name `df_fb`.
# df_yt: Imports the processed YouTube data from the file specified in `path_yt` into a DataFrame named `df_yt`.
# df_outdoor: Fetches the processed outdoor data from the file specified in `path_outdoor` into a DataFrame named `df_outdoor`.
# df_other_vars: Loads the additional variables data from the file specified in `path_other_variables` into a DataFrame named `df_other_vars`.

df_fb = pd.read_csv(path_fb)
df_yt = pd.read_csv(path_yt)
df_outdoor = pd.read_csv(path_outdoor)
df_other_vars = pd.read_csv(path_other_variables)

# This Python code prints the types of data that each column in the table called "df_other_vars" contains. It does this by using the ".dtypes" option.
# Knowing the data types in each column helps us understand how the data is stored and used.

df_other_vars.dtypes

# Here we're using Python code to change the 'week_starting_date' column in several DataFrames (df_fb, df_yt, df_outdoor, df_other_vars) into datetime format.
# This is done with the pd.to_datetime() function.
# Overall, this change makes the data ready for analysis and combining it with data from other sources based on time.
df_fb['week_starting_date'] = pd.to_datetime(df_fb['week_starting_date'])
df_yt['week_starting_date'] = pd.to_datetime(df_yt['week_starting_date'])
df_outdoor['week_starting_date'] = pd.to_datetime(df_outdoor['week_starting_date'])
df_other_vars['week_starting_date'] = pd.to_datetime(df_other_vars['week_starting_date'])

## Check date ranges of dataframes - are they your modeling timeframe?

# This Python code finds the start date of the first week and the start date of the last week in the Facebook data stored in the DataFrame `df_fb`.
# `df_fb['week_starting_date'].min()`: It finds the earliest date in the `week_starting_date` column of the DataFrame.
# - `df_fb['week_starting_date'].max()`: It finds the latest date in the `week_starting_date` column of the DataFrame.
# The printed output includes the first and last weeks of the Facebook data, providing insights into the time period covered by the dataset.

print('first week of Facebook data:' + str(df_fb['week_starting_date'].min()))
print('last week of Facebook data:' + str(df_fb['week_starting_date'].max()))

# STUDENT INPUT REQUIRED - Create similar code as above to print and check YouTube date range

print('First week of YouTube data:' + str(df_yt['week_starting_date'].min()))
print('Last week of YouTube data:' + str(df_yt['week_starting_date'].max()))

# STUDENT INPUT REQUIRED - Create similar code as above to print and check Outdoor Campaign date range

print('First week of Outdoor Campaign data:' + str(df_outdoor['week_starting_date'].min()))
print('Last week of Outdoor Campaign data:' + str(df_outdoor['week_starting_date'].max()))

# STUDENT INPUT REQUIRED - Create similar code as above to print and check All Other Variables file date range

print('First week of All Other Variables data:' + str(df_other_vars['week_starting_date'].min()))
print('Last week of All Other Variables data:' + str(df_other_vars['week_starting_date'].max()))

## Perform join of all 4 dataframes

# This Python code is intended to merge multiple DataFrames (df_fb, df_yt, df_outdoor) with the DataFrame df_other_vars based on the 'week_starting_date' column.
# Line 1 code combines the DataFrames `df_other_vars` and `df_fb` using a "left join" on the `week_starting_date` column. This means that all rows from `df_other_vars` will be present in the merged DataFrame, even if there is no matching row in `df_fb`.
# Line 2 code merges the result of Line 1 (`df_total`) with the DataFrame `df_yt` using a "left join" on the `week_starting_date` column. Again, all rows from `df_total` will be included in the merged DataFrame, regardless of matches in `df_yt`.
# Line 3 code merges the result of Line 2 (`df_total`) with the DataFrame `df_outdoor` using a "left join" on the `week_starting_date` column. This completes the process of combining multiple DataFrames based on a common date column, ensuring that all relevant information is available for analysis.

df_total = df_other_vars.merge(df_fb, on= 'week_starting_date', how='left')

df_total = df_total.merge(df_yt, on= 'week_starting_date', how='left')

df_total = df_total.merge(df_outdoor, on= 'week_starting_date', how='left')

# The following Python code is designed to perform calculations on the sum of numeric values from the previous dataframes (df_total, df_other_vars, df_fb, df_yt, df_outdoor).
# Initially, it sets the index of the df_total dataframe to the column 'week_starting_date'.
# It then calculates the total sum of numeric values in the df_total dataframe.
# By subtracting the combined sum of numeric values from the other dataframes (df_other_vars, df_fb, df_yt, df_outdoor) from this total sum.
# Essentially, it calculates the difference between the total sum of one dataframe (df_total) and the sum of values from multiple other dataframes (df_other_vars, df_fb, df_yt, df_outdoor).

df_total = df_total.set_index('week_starting_date')

# STUDENT COMMENT REQUIRED - write a short summary of what the python code in this cell is intended to do
(df_total.sum(numeric_only=True).sum() - 
          (df_other_vars.sum(numeric_only=True).sum() + 
           df_fb.sum(numeric_only=True).sum() + 
           df_yt.sum(numeric_only=True).sum() + 
           df_outdoor.sum(numeric_only=True).sum()))


## Peform check of final table

# The Python code provided here calculates the absolute difference between the total sum of numeric values in the df_total dataframe and the combined sum of numeric values from other specified dataframes like df_other_vars, df_fb, df_yt, and df_outdoor.
# After obtaining this absolute difference, it is then divided by the total sum of numeric values in the df_total dataframe.
# The resulting value is then compared to a threshold of 0.000001. If the resulting value is less than this threshold, the code prints 'Final totals match', indicating that the total sums from df_total and the other specified dataframes are very close.
# If the resulting value is greater than or equal to the threshold, it prints 'Final totals DO NOT match - revisit treatment', suggesting that there may be discrepancies in the data treatment that need to be reviewed and addressed.
# In summary, this code performs a validation check to ensure the consistency of total sums across the different dataframes mentioned.

if abs(df_total.sum(numeric_only=True).sum() - 
          (df_other_vars.sum(numeric_only=True).sum() + 
           df_fb.sum(numeric_only=True).sum() + 
           df_yt.sum(numeric_only=True).sum() + 
           df_outdoor.sum(numeric_only=True).sum()))/df_total.sum().sum() <0.000001:
    print('Final totals match')
else:
    print('Final totals DO NOT match - revisit treatment')

## Export processed data

# This Python code is intended to export the DataFrame df_total to a CSV file.
# df_total.to_csv(export_data_path + file_name): This method exports the data from the DataFrame df_total to a CSV (Comma-Separated Values) file.
# The export_data_path variable specifies the directory path where the CSV file will be saved, and the file_name variable specifies the name of the CSV file.


df_total.to_csv(export_data_path + file_name)
